refactor(engine/rocky-core): dedup compute_backoff across adapter crates
Hoist the three identical `compute_backoff` copies (rocky-core::state_sync,
rocky-databricks::connector, rocky-snowflake::connector) into a single
shared `rocky_core::retry::compute_backoff` helper. All three call sites
and their existing retry loops are unchanged; only the source of the
function moves.

Zero behaviour change. The tests from the adapter crates (exponential
without jitter, capped, with jitter in range) are consolidated into the
new module; duplicate tests in state_sync are removed.

Follow-up to #213.
