+++ engine/crates/rocky-databricks/src/connector.rs
@@ -1,8 +1,9 @@
 use rocky_core::circuit_breaker::{CircuitBreaker, TransitionOutcome};
-use std::time::{Duration, SystemTime};
+use std::time::Duration;
 
 use reqwest::Client;
 use rocky_core::config::RetryConfig;
+use rocky_core::retry::compute_backoff;
 use rocky_observe::events::{ErrorClass, PipelineEvent, global_event_bus};
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
@@ -568,27 +569,6 @@ fn poll_delay(attempt: usize) -> Duration {
     Duration::from_millis(delay_ms)
 }
 
-/// Computes backoff duration with exponential growth, capped at max, with optional jitter.
-fn compute_backoff(cfg: &RetryConfig, attempt: u32) -> u64 {
-    let base = (cfg.initial_backoff_ms as f64) * cfg.backoff_multiplier.powi(attempt as i32);
-    let capped = base.min(cfg.max_backoff_ms as f64) as u64;
-
-    if cfg.jitter {
-        // Use subsecond nanos as cheap jitter source (no rand dependency needed)
-        let jitter_range = capped / 4; // +-25% jitter
-        let nanos = SystemTime::now()
-            .duration_since(SystemTime::UNIX_EPOCH)
-            .unwrap_or_default()
-            .subsec_nanos() as u64;
-        let jitter = nanos % (jitter_range.max(1));
-        capped
-            .saturating_sub(jitter_range / 2)
-            .saturating_add(jitter)
-    } else {
-        capped
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -762,51 +742,6 @@ mod tests {
     fn test_not_transient_canceled() {
         let err = ConnectorError::Canceled { id: "id-1".into() };
         assert!(!is_transient(&err));
-    }
-
-    #[test]
-    fn test_backoff_exponential_no_jitter() {
-        let cfg = RetryConfig {
-            max_retries: 3,
-            initial_backoff_ms: 1000,
-            max_backoff_ms: 30000,
-            backoff_multiplier: 2.0,
-            jitter: false,
-            ..Default::default()
-        };
-        assert_eq!(compute_backoff(&cfg, 0), 1000); // 1000 * 2^0
-        assert_eq!(compute_backoff(&cfg, 1), 2000); // 1000 * 2^1
-        assert_eq!(compute_backoff(&cfg, 2), 4000); // 1000 * 2^2
-    }
-
-    #[test]
-    fn test_backoff_capped() {
-        let cfg = RetryConfig {
-            max_retries: 5,
-            initial_backoff_ms: 1000,
-            max_backoff_ms: 5000,
-            backoff_multiplier: 2.0,
-            jitter: false,
-            ..Default::default()
-        };
-        assert_eq!(compute_backoff(&cfg, 3), 5000); // 1000 * 2^3 = 8000, capped at 5000
-        assert_eq!(compute_backoff(&cfg, 4), 5000); // still capped
-    }
-
-    #[test]
-    fn test_backoff_with_jitter_in_range() {
-        let cfg = RetryConfig {
-            max_retries: 3,
-            initial_backoff_ms: 1000,
-            max_backoff_ms: 30000,
-            backoff_multiplier: 2.0,
-            jitter: true,
-            ..Default::default()
-        };
-        // With jitter, result should be within +-25% of base (1000)
-        let result = compute_backoff(&cfg, 0);
-        assert!(result >= 750, "backoff {result} should be >= 750");
-        assert!(result <= 1250, "backoff {result} should be <= 1250");
     }
 
     #[test]
